+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Enhancements
   * Pass the repository with the changeset to callbacks
   * Add `template`, `encoding`, `lc_collate` and `lc_ctype` options to adapter that are used when creating the database
+  * Add support for migrations (the previous migration style no longer works, just replace the SQL commands by `execute/1` calls)
 
 * Backwards incompatible changes
   * Previously deprecated validations have been removed